Silence a MSVC warning about narrowing conversion
We actually want to convert a size_t value into an int value in this
return statement.

Fix the annoying MSVC warning by actually casting the bytestride to an
int

+++ tiny_gltf.h
@@ -522,7 +522,7 @@ struct Accessor {
       if ((bufferViewObject.byteStride % componentSizeInBytes) != 0) {
         return -1;
       }
-      return bufferViewObject.byteStride;
+      return static_cast<int>(bufferViewObject.byteStride);
     }
 
     return 0;
